+++ ArchivingTestApp/MainForm.cs
@@ -1,6 +1,10 @@
 using L10NSharp;
 using SIL.Archiving;
+using SIL.Archiving.Generic;
+using SIL.Archiving.IMDI;
+using SIL.IO;
 using SIL.Windows.Forms.Archiving;
+using SIL.Windows.Forms.Archiving.IMDI;
 using static System.IO.Path;
 using static System.String;
 
@@ -9,6 +13,15 @@ namespace ArchivingTestApp
 	public partial class MainForm : Form
 	{
 		private const string kAppName = "Archiving Test App";
+
+		private string GetTitle()
+		{
+			var title = m_txtTitle.Text;
+			if (title.Length == 0)
+				title = "Arbitrary title";
+			return title;
+		}
+
 		public MainForm()
 		{
 			InitializeComponent();
@@ -16,15 +29,57 @@ namespace ArchivingTestApp
 
 		private void m_btnRamp_Click(object sender, EventArgs e)
 		{
-			var title = m_txtTitle.Text;
-			if (title.Length == 0)
-				title = "Arbitrary title";
+			var title = GetTitle();
 			var model = new RampArchivingDlgViewModel(kAppName, title,
 				title.ToLatinOnly("~", "_", ""), SetFilesToArchive, GetFileDescription);
 			using (var rampArchiveDlg = new ArchivingDlg(model, LocalizationManager.GetString(
 				"ArchivingTestApp.MainForm.AdditionalArchiveProcessInfo", "This is just a test.")))
 			{
 				rampArchiveDlg.ShowDialog(this);
+			}
+		}
+
+		private void m_btnIMDI_Click(object sender, EventArgs e)
+		{
+			var title = GetTitle();
+			TempFile.NamePrefix = kAppName.Replace(" ", "_");
+			var folder = new TempFile().Path;
+			RobustFile.Delete(folder);
+			Directory.CreateDirectory(folder);
+			var model = new IMDIArchivingDlgViewModel(kAppName, title,
+				title.ToLatinOnly("~", "_", ""), false, SetFilesToArchive, folder);
+
+			model.ArchivingPackage.AccessCode = "internal";
+			model.ArchivingPackage.Access.DateAvailable = "3 March 2029";
+			model.ArchivingPackage.Access.Owner = "Fred";
+			model.ArchivingPackage.Publisher = "SIL Elbonia";
+			model.ArchivingPackage.Location = new ArchivingLocation
+			{
+				Address = "1234 Shoulder St.; Armstrong",
+				Country = "Elbonia"
+			};
+			model.ArchivingPackage.Author = "Test Dude";
+			model.ArchivingPackage.Owner = "Mike";
+
+			foreach (ListViewGroup group in m_listFiles.Groups)
+			{
+				var session = model.AddSession(group.Header);
+				foreach (var file in (from ListViewItem item in @group.Items select item.Text))
+				{
+					session.AddFile(new ArchivingFile(file));
+					session.AddFileAccess(file, (ArchivingPackage)model.ArchivingPackage);
+				}
+
+				session.Genre = "Dance";
+				session.SubGenre = "Entertainment";
+				session.PlanningType = "Spontaneous";
+			}
+
+
+			using (var imdiArchiveDlg = new IMDIArchivingDlg(model, LocalizationManager.GetString(
+				       "ArchivingTestApp.MainForm.AdditionalImdiArchiveProcessInfo", "This is a test of IMDI archival.")))
+			{
+				imdiArchiveDlg.ShowDialog(this);
 			}
 		}
 
@@ -41,7 +96,7 @@ namespace ArchivingTestApp
 
 		private static string GetFileDescription(string key, string filename)
 		{
-			return $"{key} - {filename}";
+			return filename.Replace("\\", "_");
 		}
 
 		private void HandleAddFilesClick(object sender, EventArgs e)